POST Request Method to Delete the Article

+++ delete_article.php
@@ -27,21 +27,23 @@ if(isset($_GET['id'])) {
 
 }
 
-$sql = "DELETE FROM article
-        WHERE id = ?;"; 
-        
-$stmt = mysqli_prepare($conn, $sql);
+if($_SERVER["REQUEST_METHOD"] == "POST") {
+    $sql = "DELETE FROM article
+    WHERE id = ?;"; 
+    
+    $stmt = mysqli_prepare($conn, $sql);
 
-if ($stmt === false) {
-    echo mysqli_error($conn);
-} else {
+    if ($stmt === false) {
+        echo mysqli_error($conn);
+    } else {
 
     mysqli_stmt_bind_param($stmt, "i", $id);
-            
+        
     if(mysqli_stmt_execute($stmt)) {
 
         redirect("/index.php");
     } else {
         echo mysqli_stmt_error($stmt);
+    }
     }
 }
